+++ builtin/blame.c
@@ -672,7 +672,7 @@ static void sanity_check_on_fail(struct blame_scoreboard *sb, int baa)
 static unsigned parse_score(const char *arg)
 {
 	char *end;
-	unsigned long score = strtoul(arg, &end, 10);
+	size_t score = strtoul(arg, &end, 10);
 	if (*end)
 		return 0;
 	return score;
