+++ builtin/pack-redundant.c
@@ -629,7 +629,7 @@ int cmd_pack_redundant(int argc, const char **argv, const char *prefix)
 
 	if (verbose) {
 		fprintf(stderr, "There are %lu packs available in alt-odbs.\n",
-			(unsigned long)pack_list_size(altodb_packs));
+			pack_list_size(altodb_packs));
 		fprintf(stderr, "The smallest (bytewise) set of packs is:\n");
 		pl = min;
 		while (pl) {
@@ -638,10 +638,10 @@ int cmd_pack_redundant(int argc, const char **argv, const char *prefix)
 		}
 		fprintf(stderr, "containing %lu duplicate objects "
 				"with a total size of %lukb.\n",
-			(unsigned long)get_pack_redundancy(min),
-			(unsigned long)pack_set_bytecount(min)/1024);
+			get_pack_redundancy(min),
+			(size_t)pack_set_bytecount(min)/1024);
 		fprintf(stderr, "A total of %lu unique objects were considered.\n",
-			(unsigned long)all_objects->size);
+			(size_t)all_objects->size);
 		fprintf(stderr, "Redundant packs (with indexes):\n");
 	}
 	pl = red = pack_list_difference(local_packs, min);
@@ -653,7 +653,7 @@ int cmd_pack_redundant(int argc, const char **argv, const char *prefix)
 	}
 	if (verbose)
 		fprintf(stderr, "%luMB of redundant packs in total.\n",
-			(unsigned long)pack_set_bytecount(red)/(1024*1024));
+			(size_t)pack_set_bytecount(red)/(1024*1024));
 
 	return 0;
 }
