+++ builtin/update-index.c
@@ -504,7 +504,7 @@ static void read_index_info(int nul_term_line)
 		char *path_name;
 		struct object_id oid;
 		unsigned int mode;
-		unsigned long ul;
+		size_t ul;
 		int stage;
 
 		/* This reads lines formatted in one of three formats:
@@ -831,7 +831,7 @@ static int parse_new_style_cacheinfo(const char *arg,
 				     struct object_id *oid,
 				     const char **path)
 {
-	unsigned long ul;
+	size_t ul;
 	char *endp;
 	const char *p;
